+++ src/util/mod.rs
@@ -4,6 +4,7 @@ pub mod fs;
 pub mod hashing;
 pub mod network;
 pub mod path;
+pub mod path_entity;
 pub mod process;
 pub mod query;
 pub mod term;